backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
django-environ==0.11.2

# AI/LLM Dependencies
langchain-groq==0.1.9
langchain-core==0.3.15
langchain==0.3.7